application functionnal but Bob do not successfuly get the correct answer

+++ src/utils/Key.java
@@ -1,6 +1,8 @@
 package utils;
 
+import java.math.BigInteger;
+
 public abstract class Key {
 	
-	protected int n;
+	protected BigInteger n;
 }
